requirements.txt
Django>=1.6,<1.9
isodate>=0.4.7,<0.6
django-extensions==1.7.4
django-enumfields==0.8.2
pyyaml==3.12
django-rq==0.9.2
requests==2.11.1
docker-py==1.10.3
dj-database-url==0.4.1
psycopg2==2.6.2
